+++ apps/api/src/app/portfolio/interfaces/portfolio-position-detail.interface.ts
@@ -13,6 +13,7 @@ export interface PortfolioPositionDetail {
   minPrice: number;
   quantity: number;
   symbol: string;
+  transactionCount: number;
 }
 
 export interface HistoricalDataItem {