+++ apps/api/src/app/portfolio/interfaces/portfolio-position.interface.ts
@@ -20,6 +20,7 @@ export interface PortfolioPosition {
   sector?: string;
   shareCurrent: number;
   shareInvestment: number;
+  transactionCount: number;
   symbol: string;
   type?: string;
   url?: string;